Added table checking to existing database migrations.

+++ partEZ/database/migrations/2014_10_12_000000_create_users_table.php
@@ -12,15 +12,19 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->increments('uid')->unsigned();
-            $table->string('firstname');
-            $table->string('lastname');
-            $table->string('email')->unique();
-            $table->string('password', 60);
-            $table->rememberToken();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users'))
+        {
+            Schema::create('users', function (Blueprint $table) {
+                $table->increments('uid')->unsigned();
+                $table->string('firstname');
+                $table->string('lastname');
+                $table->string('email')->unique();
+                $table->string('password', 60);
+                $table->rememberToken();
+                $table->timestamps();
+            });
+        }
+        
     }
 
     /**
@@ -30,6 +34,9 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('users');
+        if (Schema::hasTable('users'))
+        {
+            Schema::drop('users');
+        }
     }
 }
